Main installation fail message

+++ installAll.sh
@@ -34,6 +34,12 @@ done
 
 cd Main0
 ghc Main.hs
+if [[ $? -ne 0]]
+then
+	echo "Installing Main failed"
+	exit
+fi
+
 mv Main ../Main
 rm *.hi
 rm *.o
